media-source/MediaSource-duration.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>

<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }        
        processBody(new Uint8Array(xhr.response));
    };
}
test(function() {
    var mediaSource = new MediaSource();
    assert_equals(mediaSource.duration, NaN);
    assert_throws({name : 'InvalidAccessError'},
                  function() {
                      mediaSource.duration = -1;
                  }, 'Set invalid duration');
    assert_throws({name : 'InvalidAccessError'},
                  function() {
                      mediaSource.duration = NaN;
                  }, 'Set invalid duration');

    assert_throws({name : 'InvalidStateError'},
                  function() {
                      mediaSource.duration = 1;
                  }, 'Not opened yet');
}, 'MediaSource#duration : initial state');

async_test(function(t) {
    GET(FILE, function(uInt8Array) {
        var data = new Blob([uInt8Array], {type: 'video/webm'});
        var chunkNum = 5;
        var chunkSize = Math.ceil(data.size / chunkNum);
        var chunkIdx = 0;
        var mediaSource = new MediaSource();
        
        mediaSource.addEventListener('sourceopen', function(e){
            //playback 1st chunk
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            var startByte = chunkSize * chunkIdx;
            var chunk = data.slice(startByte, startByte + chunkSize);
            var reader = new FileReader();
            reader.onload = function(e) {
                var data = new Uint8Array(e.target.result);
                sourceBuffer.appendBuffer(data);
                chunkIdx++;
            };
            reader.readAsArrayBuffer(chunk);

            sourceBuffer.addEventListener('update', function(e) {                
                t.step(function() {
                    assert_throws({name : 'InvalidStateErrors'},
                                  function() { mediaSource.duration = 1; },
                                  'Set invalid duration');
                });
                t.done();
            });
        }, false);
        video.src = window.URL.createObjectURL(mediaSource);
    });
}, 'MediaSource#duration : update during SourceBuffer#updating is true');

</script>
</body>
</html>
